adafruit should install arduino samd

+++ arduino.js
@@ -83,7 +83,9 @@ function initTasks(gulp, options) {
   });
 
   gulp.task('install-tools-package', false, function (cb) {
-    installPackage(options.board.package, options.board.arch, options.board.packageUrl, cb);
+    installPackage('arduino', 'samd', null, function() {
+      installPackage(options.board.package, options.board.arch, options.board.packageUrl, cb);
+    });
   })
 
   gulp.task('install-tools-libraries', false, function (cb) {
@@ -280,7 +282,7 @@ function installLibraries(libs, cb) {
 function installPackage(name, arch, addUrl, cb) {
 
   // make sure package index exists, if it doesn't exist, try to clean up directory to make sure no uncomplete installation exists
-  if (!all.fileExistsSync(getArduino15Folder() + '/' + addUrl.split('/').slice(-1)[0])) {
+  if (addUrl && !all.fileExistsSync(getArduino15Folder() + '/' + addUrl.split('/').slice(-1)[0])) {
     all.deleteFolderRecursivelySync(getPackageFolder() + '/' + name);
   }
